データベースがMariaDBの場合は全文検索機能を無効化

+++ database/migrations/2020_10_24_161242_add_fulltext_index_to_pages.php
@@ -1,5 +1,6 @@
 <?php
 
+use App\Eloquents\Page;
 use Illuminate\Database\Migrations\Migration;
 use Illuminate\Database\Schema\Blueprint;
 use Illuminate\Support\Facades\DB;
@@ -13,7 +14,9 @@ class AddFulltextIndexToPages extends Migration
      */
     public function up()
     {
-        DB::statement('ALTER TABLE pages ADD FULLTEXT INDEX fulltext_index (title,body) WITH PARSER ngram');
+        if (Page::isMySqlFulltextIndexSupported()) {
+            DB::statement('ALTER TABLE pages ADD FULLTEXT INDEX fulltext_index (title,body) WITH PARSER ngram');
+        }
     }
 
     /**
@@ -23,6 +26,8 @@ class AddFulltextIndexToPages extends Migration
      */
     public function down()
     {
-        DB::statement('ALTER TABLE pages DROP INDEX fulltext_index');
+        if (Page::isMySqlFulltextIndexSupported()) {
+            DB::statement('ALTER TABLE pages DROP INDEX fulltext_index');
+        }
     }
 }
